+++ docs/release-notes.md
@@ -14,6 +14,7 @@ nav_order: 9
 - Ship aarch64 macOS ignition-validate binary in GitHub release artifacts
 - Allow enabling discard passthrough on LUKS devices _(3.4.0-exp)_
 - Allow specifying arbitrary LUKS open options _(3.4.0-exp)_
+- Support offline Tang provisioning via pre-shared advertisement _(3.4.0-exp)_
 
 ### Changes